Reimplement in JavaScript
I don't particularly like it, but this implementation is about 2x
faster. This is partly due to just regular PureScript overhead, partly
due to an algorithmic change: we now keep keys, values, and subnodes
inline in nodes, in an heterogeneous `content` array, as described by
Steindorfer & Vinju (OOPSLA 2015). This could have been done in
PureScript, with liberal use of unsafeCoerce, but it's not much
prettier.

This also puts Data.Hashable into this library. We can always pull it
out later again.