These TODOs are actually unreachable
I never need to union a regular node with a collision node, or the
other way round, because collision nodes are always at maximum depth.
I think in theory that is not necessary, we could have collisions
earlier in the tree. However, collisions are terrible anyway, so it is
not a big loss.